photo1

+++ include/high_level.h
@@ -11,7 +11,7 @@ struct node
     std::vector<std::vector<std::vector<int > > > cts_point;// cts_point[i]  [j] = (x,y,t)
     std::vector<std::vector<std::vector<int> > > cts_edge;// cts_edge[i] [j] = (x1,y2,t1,x2,y2,t2)
     std::vector<std::vector<std::vector<int > > > phs;
-    int cost;
+    double cost;
     void set(agent* as);
 
     node inherit();
@@ -31,5 +31,5 @@ void split(node root,std::vector<int > ct_now,std::priority_queue<node,std::vect
 
 std::vector<int > get_ct_point5(std::vector<std::vector<int > > cts_pointbig,node nnow,bool PC=false);
 
-int runCBS(agent* as,int n,bool pc=false);
+double runCBS(agent* as,int n,bool pc=false);
 #endif
